EngineFixes compatibility. Squashed commits for readability, so most recent update comments come first.
Case independent comparison throughout. Check if config is the same, rather than just a signature.

Fix numeric replace specify when followed by digit. $18192-> $018192. Several text clarity tweaks

Make EF match STR's MaxStdio of 8192.
Engine Fixes compatibility.
Updates squashed for readability, so comments are most-reccent first.

EngineFixes compatibility.
Squashed the commits for readability, so most-recent comment is first.

Case-independent regex matching.

Enable compatibility with EngineFixes.

Fixed incomplete refactor, a couple of cleanups.

Just a few lines in ScriptExtender.cpp to load the DLLs in the right order.

But, EF isn't compatible if some features are turned on, notoriously tbbmalloc. The bulk of the change is in DllBlocklist, to detect bad configurations of the mod and fix them. It's general-purpose enough to use with other mods if / when that comes up.

Would have been a lot simpler but for TOML configuration rules.