Prevent multiple triggers of quest stage change
This started as a bug hunt for misc. quest syncing, but it effects ALL quest type progression. When the quest stage is set, syncing will cause each player to try to set it, which can cause duplication of steps or unwanted progression.

The fix is to make sure desired quest stage != current quest stage.