Fix logic that determines whether to spoof GetModuleFileNameW(nullptr)
It was depending on figuring out where it was called from, but the logic was unreliable and there is no apparent way to fix it. Now, it switches spoofing on just before loading SkyrimSE.exe (so the loaded DLLs see the path they expect).

Requires a companion fix in TIltedCore:TiltedPhoques::GetPath() to persist the value found originally, so it doesn't change out from under with late callers. If you code has only late callers, just add a call before loading your game binary.